backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2==2.9.6
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.0